remote-communicationのmissionのアダプター構成をリファクタリング
