+++ remote-communication/pkg/mission/adapters/grpc/upload_mission_edge_service_server_test.go
@@ -1,4 +1,4 @@
-package ports
+package grpc
 
 import (
 	"remote-communication/pkg/mission/app"
