+++ remote-communication/pkg/mission/adapters/rabbitmq/copied_mission_created_event_handler.go
@@ -1,4 +1,4 @@
-package ports
+package rabbitmq
 
 import (
 	"context"
