+++ remote-communication/pkg/mission/adapters/rabbitmq/copied_mission_created_event_handler_test.go
@@ -1,4 +1,4 @@
-package ports
+package rabbitmq
 
 import (
 	"remote-communication/pkg/mission/app"
