+++ remote-communication/pkg/mission/adapters/rabbitmq/subscribe_event_handler_test.go
@@ -1,4 +1,4 @@
-package ports
+package rabbitmq
 
 import (
 	"context"
